support 32-bit hosts in kernel/posix/time

+++ go/kernel/posix/time.go
@@ -12,7 +12,7 @@ func (k *PosixKernel) ClockGettime(_ int, out co.Obuf) uint64 {
 	var err error
 	ts := syscall.NsecToTimespec(time.Now().UnixNano())
 	if k.U.Bits() == 64 {
-		err = out.Pack(&native.Timespec64{ts.Sec, ts.Nsec})
+		err = out.Pack(&native.Timespec64{int64(ts.Sec), int64(ts.Nsec)})
 	} else {
 		err = out.Pack(&native.Timespec{int32(ts.Sec), int32(ts.Nsec)})
 	}
